+++ changelog.md
@@ -1,4 +1,13 @@
 # Changelog
 
-**./src/output/file-write/graph-steps/layout-definition.js**
-* Added missing comments to file.
+**Edge Highlighting**
+* Now applied properly for Cytoscape graphs.
+* The glitch was that the precompiled graph's edges were not highlighted properly.
+* Mapped graph files appear to have been unaffected.
+
+---
+
+**./src/pathfinding/route-precompiled.js**
+* Modified 'edgeStatus.setHighlights' call arguments.
+	* Before: 'compiledObject.graphData'
+	* After: 'compiledObject.graphData.edgeList'